+++ src/BugsnagPerformance/BugsnagPerformance.sln
@@ -5,7 +5,9 @@ VisualStudioVersion = 16.0.810.23
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BugsnagPerformance", "BugsnagPerformance\BugsnagPerformance.csproj", "{94AA2CBE-D78E-427A-B587-FB5AA96A2063}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BugsnagPerformance.Tests", "BugsnagPerformance.Tests\BugsnagPerformance.Tests.csproj", "{5BC86E8D-9A24-44E6-A30B-66751EA1FAF5}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UnityEngine", "..\..\unitTests\UnityEngine\UnityEngine.csproj", "{26C1DA2F-681B-4544-8E8E-03B6E2F9B9D8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BugsnagPerformance.Tests", "..\..\unitTests\BugsnagPerformance.Tests\BugsnagPerformance.Tests.csproj", "{6410C645-AB95-41D4-9BC8-574182A23B81}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -17,10 +19,14 @@ Global
 		{94AA2CBE-D78E-427A-B587-FB5AA96A2063}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{94AA2CBE-D78E-427A-B587-FB5AA96A2063}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{94AA2CBE-D78E-427A-B587-FB5AA96A2063}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5BC86E8D-9A24-44E6-A30B-66751EA1FAF5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5BC86E8D-9A24-44E6-A30B-66751EA1FAF5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5BC86E8D-9A24-44E6-A30B-66751EA1FAF5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5BC86E8D-9A24-44E6-A30B-66751EA1FAF5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{26C1DA2F-681B-4544-8E8E-03B6E2F9B9D8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{26C1DA2F-681B-4544-8E8E-03B6E2F9B9D8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{26C1DA2F-681B-4544-8E8E-03B6E2F9B9D8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{26C1DA2F-681B-4544-8E8E-03B6E2F9B9D8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6410C645-AB95-41D4-9BC8-574182A23B81}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6410C645-AB95-41D4-9BC8-574182A23B81}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6410C645-AB95-41D4-9BC8-574182A23B81}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6410C645-AB95-41D4-9BC8-574182A23B81}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
