+++ unitTests/BugsnagPerformance.Tests/Properties/AssemblyInfo.cs
