+++ unitTests/BugsnagPerformance.Tests/packages.config
