+++ unitTests/UnityEngine/Debug.cs
@@ -1,0 +1,14 @@
+﻿using System;
+namespace UnityEngine
+{
+    public class Debug
+    {
+        public Debug()
+        {
+        }
+
+        public static void LogWarning(object obj)
+        { }
+
+    }
+}
