+++ unitTests/UnityEngine/MonoBehaviour.cs
@@ -1,0 +1,10 @@
+﻿using System;
+namespace UnityEngine
+{
+    public class MonoBehaviour
+    {
+        public MonoBehaviour()
+        {
+        }
+    }
+}
