+++ unitTests/UnityEngine/UnityEngine.csproj.user
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup Condition=" '$(RunConfiguration)' == 'Default' ">
+    <StartAction>Project</StartAction>
+    <ConsolePause>true</ConsolePause>
+  </PropertyGroup>
+</Project>
